Moving assets to client folder
