+++ cachiql-demo/src/client/Components/AboutUs.tsx
@@ -6,11 +6,11 @@ import Container from '@material-ui/core/Container';
 import { makeStyles } from '@material-ui/core/styles';
 import IconButton from '@material-ui/core/IconButton';
 import Typography from '@material-ui/core/Typography';
-import Kaden from '../../assets/resizedKaden-svg.svg';
-import Vanessa from '../../assets/resizedVanessa-svg.svg';
-import Eddy from '../../assets/resizedEddy-svg.svg';
-import Fahad from '../../assets/resizedFahad-svg.svg';
-import TeamCachiql from '../../assets/teamCachiql.svg';
+import Kaden from '../assets/resizedKaden-svg.svg';
+import Vanessa from '../assets/resizedVanessa-svg.svg';
+import Eddy from '../assets/resizedEddy-svg.svg';
+import Fahad from '../assets/resizedFahad-svg.svg';
+import TeamCachiql from '../assets/teamCachiql.svg';
 import { Grid } from '@material-ui/core';
 
 const useStyles = makeStyles((theme) => ({
